Somehow getting an NPE in main, a biome not being assigned weird

+++ src/main/java/com/azortis/orbis/biomedemo/objects/Region.java
@@ -45,7 +45,7 @@ public class Region {
     private long seed;
 
     // The zoom
-    private double zoom;
+    private int zoom;
 
     private List<RegionLayer> landRegions;
     private List<BiomeLayer> landBiomes;
@@ -73,7 +73,7 @@ public class Region {
         return seed;
     }
 
-    public double getZoom() {
+    public int getZoom() {
         return zoom;
     }
